requirements.txt
Flask
flask_cors
flask_pymongo
python-dotenv
dnspython
werkzeug
flask-session
pytz
requests
pandas
pyspark
